requirements.txt
## The following requirements were added by pip freeze:
Django==2.0
django-appconf==1.0.2
django-bootstrap-form==3.3
django-user-accounts==2.0.3
pinax-templates==1.0.4
pinax-theme-bootstrap==8.0.1
pytz==2017.3
